extract pool to separate file
