+++ crates/monty-type-checking/src/lib.rs
@@ -1,4 +1,5 @@
 mod db;
+mod pool;
 mod type_check;
 
 pub use crate::type_check::{SourceFile, TypeCheckingDiagnostics, type_check};
